- Add razz color as primary color. - toDouble() conversion for screenUtils.

+++ packages/stream_chat_flutter/lib/src/channel_header.dart
@@ -210,7 +210,7 @@ class ChannelHeader extends StatelessWidget implements PreferredSizeWidget {
                     const SizedBox(width: 6),
                     title ?? ChannelName(
                         textStyle: channelHeaderTheme.titleStyle!.copyWith(
-                            fontSize: ScreenUtil().setSp(16.0),
+                            fontSize: ScreenUtil().setSp(16.0).toDouble(),
                             fontFamily: 'Poppins'
                         ),
                       ),
